% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmr_dat}
\alias{cmr_dat}
\title{cmr_dat}
\format{A data frame with five variables: \code{id}, \code{loc},
  \code{x}, \code{y} and \code{date}. \code{cmr_dat} is an example of how dataframes should be created to feed into CMRdat.}
\usage{
cmr_dat
}
\description{
cmr_dat
}
\keyword{datasets}
