% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmr_dat2}
\alias{cmr_dat2}
\title{cmr_dat2}
\format{A data frame with six variables: \code{id}, \code{loc},
  \code{x}, \code{y}, \code{date} and \code{layer}. \code{cmr_dat2} is an example of how dataframes should be created to feed into CMRdat when there are multiple layered networks (i.e. a separate network for males and females).}
\usage{
cmr_dat2
}
\description{
cmr_dat2
}
\keyword{datasets}
