% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrData}
\alias{cmrData}
\title{cmrData
cmrData}
\format{A data frame with five variables: \code{id}, \code{loc},
  \code{x}, \code{y} and \code{date}. \code{cmrData} is an example of how dataframe es should be created to feed into CMRnet.}
\usage{
cmrData
}
\description{
cmrData can be used to test social and movement ntwork functions of CMRnet. It contains five columns: the ID column provides the identity of the individual captured, loc is the named capture location, x and y provide the geographic location of a capture event, and date provide the capture time.
}
\keyword{datasets}
