% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrData2}
\alias{cmrData2}
\title{cmrData2}
\format{
A data frame with 1191 rows and 5 columns variables:
\itemize{
\item \code{id} - provides the identity of the individual captured
\item \code{loc} - the named capture location
\item \code{x} - geographic location of a capture event (on the horizontal axis)
\item \code{y} - geographic location of a capture event (on the vertical axis)
\item \code{date} - the capture date
\item \code{layer} - provides information used to define the layers of the multiplex network
}
}
\usage{
data("cmrData2")
}
\description{
A dataset that tests \code{\link{MultiMoveNetCreate}}. \code{cmrData2} is an example of how dataframes should be created to feed into CMRnet when there are multiple layered movement networks (i.e. a separate network for males and females).
}
\keyword{datasets}
