% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrData3}
\alias{cmrData3}
\title{cmrData3}
\format{
A data frame with 1191 rows and 5 columns variables:
\itemize{
\item \code{id} - provides the identity of the individual captured
\item \code{loc} - the named capture location
\item \code{x} - geographic location of a capture event (on the horizontal axis)
\item \code{y} - geographic location of a capture event (on the vertical axis)
\item \code{date} - the capture date in POSIXct format
}
}
\usage{
data("cmrData3")
}
\description{
cmrData3 to test \code{\link{DynamicNetCreateHi}} and \code{\link{MoveNetCreateHi}}. These functions help create short-term social and movement networks.\code{cmrData3} is an example of how dataframes should be created to feed into CMRnet when capture times are relevant as well as capture dates.
}
\keyword{datasets}
