% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmrMultiplexPlot.R
\name{cmrMultiplexPlot}
\alias{cmrMultiplexPlot}
\title{cmrMultiplexPlot
#'
igraph plotting for CMRnet object}
\usage{
cmrMultiplexPlot(nets, fixed_locs = c(TRUE, FALSE), locs = NULL,
  dynamic = c(TRUE, FALSE), rows = 3, layer_colours)
}
\arguments{
\item{nets}{A list of multiplex igraph networks output by cmr_igraph}

\item{fixed_locs}{(TRUE/FALSE). Whether to keep the the locations of the nodes fixed between network windows or allow them to change}

\item{locs}{A matrix or dataframe giving x and y coordinates of all nodes in the full network. Can be used to provide user-defined layouts to the graph - helpful for plotting locations in geographic space}

\item{dynamic}{(TRUE/FALSE). If true this function will plot networks for the different network windows sequentially in the same plot window with a delay. If false then this function will plot the networks in a single, multipanel figure}

\item{rows}{Sets the number of rows in the multipanel figure is dynamic=FALSE}

\item{layer_colours}{Set the colour of the plane and nodes for each layer of the multiplex network. Needs to be a vector the same length as the number of layers in the network}
}
\value{
A series of network plots showing the multiplex movement network from each network window
}
\description{
cmrMultiplexPlot
#'
igraph plotting for CMRnet object
}
\details{
This function plots multiplex movement networks produced by CMRnet. It is developmental and it has fairly basic functionality currently. Networks can be plotted sequentially (dynamic=TRUE) or in a single multi-panelled figure (dynamic=FALSE)
}
\examples{
\dontrun{

# load in data
data(cmrData2)

# set parameters
mindate<-"2010-01-01"
maxdate<-"2015-01-01"
intwindow<-60
netwindow<-12
overlap<-0

multimovenetdat <-
 MultiMoveNetCreate(
   data = cmrData2,
   intwindow = intwindow,
   mindate = mindate,
   maxdate = maxdate,
   netwindow = netwindow,
   overlap = overlap,
   nextonly = TRUE
 )

# convert to networks
networks<-cmr_igraph(movenetdat,type="movement")
cmrMultiplexPlot(nets=networks,
                fixed_locs=TRUE,
                dynamic=FALSE,
                rows=2,
                layer_colours=c("firebrick","dodgerblue"))
}
}
