% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmrPermSpat.R
\name{cmrPermSpat}
\alias{cmrPermSpat}
\title{cmrPermSpat}
\usage{
cmrPermSpat(D, locmat, same.time, time.restrict, spat.restrict, same.id,
  n.swaps, n.rand, burnin, n.burnin, warn.thresh)
}
\arguments{
\item{D}{The input dataset to be randomised}

\item{locmat}{A matrix indicating whether distance between the capture locations is less than spat.restrict if not "n"}

\item{same.time}{(TRUE/FALSE) Whether swaps should be restricted to only occur trapping events on the same date or not}

\item{time.restrict}{Provided as a number of months. Imposes time restrictions on when swaps can take place so that locations can only be swapped with those a fixed time before or after being captured}

\item{spat.restrict}{Provided on the same scale as the coordinates in the input dataset. Imposes space restrictions on when swaps can take place so that locations can only be swapped with those captued within a fixed distance}

\item{same.id}{(TRUE/FALSE) Whether swaps should be restricted to only be between captures of the same individual}

\item{n.swaps}{The number of swaps between each random network being extracted (e.g. n.swaps = 10 would equate to 10 swaps taking place between each random network being saved)}

\item{n.rand}{The number of randomised networks to be generated}

\item{burnin}{(TRUE/FALSE) Whether burnin is required}

\item{n.burnin}{The number of swaps to discard as burn-in before the first random network is created. The total number of swaps conducted is thus n.burnin+n.swaps*n.rand}

\item{warn.thresh}{The number of times no matches are found (i.e. constraints on randomisations are too restrictive) before the function is stopped and an error message returned}
}
\value{
A randomised dataset with the same dimensions as the original input dataset
}
\description{
An internal function that operate within CMRnet::DatastreamPermSpat
}
