% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicNetCreateHi.R
\name{DynamicNetCreateHi}
\alias{DynamicNetCreateHi}
\title{DynamicNetCreateHi}
\usage{
DynamicNetCreateHi(data, intwindow, mindate, maxdate, netwindow, overlap,
  spacewindow, index = FALSE)
}
\arguments{
\item{data}{A 5 column dataframe with columns for the ID of the captured individual, the location of its capture (a name or number), the x coordinate of its capture location, the y coordinate of the capture location, and the date and time of capture}

\item{intwindow}{The maximum period of time (in minutes) between two co-captures (i.e. if \code{intwindow = 60} then two individuals captured 60 minutes apart could be considered co-captured but two individuals captured 61 days apart couldn't)}

\item{mindate}{The start date (format = \code{"YYYY-MM-DD hh:mm:ss"}) of the study (i.e. when you want to build networks from)}

\item{maxdate}{The end date (format = \code{"YYYY-MM-DD hh:mm:ss"}) of the study (i.e. when you want to build networks until). Please provide as the day after the last day of the study.}

\item{netwindow}{The period of time over which each network is built in days (i.e. \code{netwindow=30} would correspond to monthly networks)}

\item{overlap}{The amount of overlap between netwindows in days (i.e. \code{overlap = 5} would result in a second network window starting 5 days before the end of the first). When \code{overlap=0}, there is no overlap between successive network windows}

\item{spacewindow}{The maximum distance between locations that can be classed as a co-capture (calculated using the coordinate system provided in in the input data). Best used when multiple capture locations occur very close together}

\item{index}{Defaults to FALSE. If FALSE edges are weighted by the number of co-captures. If TRUE then edges are weighted by the number by an association index (no. co-captures / (total captures of A + total captures of B - no. of captures))}
}
\value{
A list of length 3 containing: \cr
\enumerate{
   \item The edgelist for the social network in each of the netwindows as an array
   \item The adjacency matrix for the social network in each of the netwindows as an array
   \item A matrix indicating which individuals occurred in each netwindow
}
}
\description{
This function creates dynamic social networks from capture-mark-recapture datasets using information on co-capture
The time period for each network, together with the temporal and spatial restrictions on the co-capture window can be defined by the user. This function includes times and dates and can be used for data with higher temporal resolution than DynamicNetCreate()
}
\examples{
\dontrun{
# load example data
data(cmrData3)

# set parameters for network creation
mindate<-"2019-12-01 00:00:00"
maxdate<-"2020-07-01 00:00:00"
intwindow<-24*60
netwindow<-20
overlap<-2
spacewindow<-0

# create dynamic network
netdat<-DynamicNetCreate(data=cmrData,
intwindow=intwindow,
mindate=mindate,
maxdate=maxdate,
netwindow=netwindow,
overlap=overlap,
spacewindow=spacewindow,
index=FALSE)
}
}
