% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveNetCreateHi.R
\name{MoveNetCreateHi}
\alias{MoveNetCreateHi}
\title{MoveNetCreate}
\usage{
MoveNetCreateHi(data, intwindow, mindate, maxdate, netwindow, overlap,
  nextonly = FALSE, index = FALSE)
}
\arguments{
\item{data}{A 5 column dataframe with columns for the ID of the captured individual, the location of its capture (a name or number), the x coordinate of its capture location, the y coordinate of the capture location, and the date and time of capture}

\item{intwindow}{The maximum period of time (in minutes) between two co-captures (i.e. if \code{intwindow = 60} then two individuals captured 60 minutes apart could be considered co-captured but two individuals captured 61 days apart couldn't)}

\item{mindate}{The start date (format = \code{"YYYY-MM-DD hh:mm:ss"}) of the study (i.e. when you want to build networks from)}

\item{maxdate}{The end date (format = \code{"YYYY-MM-DD hh:mm:ss"}) of the study (i.e. when you want to build networks until). Please provide as the day after the last day of the study. Please provide as the day after the last day of the study.}

\item{netwindow}{The period of time over which each network is built in days (i.e. \code{netwindow=30} would correspond to monthly networks)}

\item{overlap}{The amount of overlap between netwindows in days (i.e. \code{overlap = 5} would result in a second network window starting 5 days before the end of the first). When \code{overlap=0}, there is no overlap between successive network windows}

\item{nextonly}{(TRUE/FALSE). Determines whether a network edge is only created to the next capture of an individual or all captures within the intwindow. Defaults to FALSE}

\item{index}{Defaults to FALSE. If FALSE edges are weighted by the number of movements. If TRUE then edges are weighted by the number of movements divided by the number of captures in a group}
}
\value{
A list of length 3 containing:
\enumerate{
   \item the edgelist for the network in each of the netwindows as an array
   \item the adjacency matrix for the network in each of the netwindows as an array
   \item a matrix indicating which individuals occurred in each netwindow
}
}
\description{
This function creates dynamic, directed movement networks from capture-mark-recapture datasets using information on the capture locations and times of individuals. Networks connect locations that individuals have moved between within a particular interaction window. The time period for each network, together with the temporal and spatial restrictions on the capture window used to infer a movement can be defined by the user
}
\examples{
\dontrun{
# load example data
data(cmrData3)

# set parameters for network creation
mindate<-"2019-12-01 00:00:00"
maxdate<-"2020-07-01 00:00:00"
intwindow<-24*60
netwindow<-20
overlap<-2
spacewindow<-0

# create network
movenetdat<-MoveNetCreateHi(data=cmrData,
intwindow=intwindow,
mindate=mindate,
maxdate=maxdate,
netwindow=netwindow,
overlap=overlap,
nextonly=TRUE,
index=FALSE)
}
}
