% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplexMoveNetCr.R
\name{MultiMoveNetCreate}
\alias{MultiMoveNetCreate}
\title{MultiMoveNetCreate}
\usage{
MultiMoveNetCreate(data, intwindow, mindate, maxdate, netwindow, overlap,
  nextonly = FALSE)
}
\arguments{
\item{data}{A 6 column dataframe with columns for the ID of the captured individual, the location of its capture (a name or number), the x coordinate of its capture location, the y coordinate of the capture location, the date of capture and infomration to define multiplex network layers. The "Layers" column can consist of any unique identifiers (e.g. if layers representing movements by males and females are used then they could be represented by "M" and "F" or 1 and 2). If the user wants a layer per individual then the "Layers" column can simply be a copy of the individual ID column/}

\item{intwindow}{The maximum period of time (in days) between two co-captures (i.e. if intwindow = 10 then two individuals captured 10 days apart could be considered co-captured but two indivviduals captured 11 days apart couldn't)}

\item{mindate}{The start date ("YYYY-MM-DD") of the study (i.e. when you want to build networks from)}

\item{maxdate}{The end date ("YYYY-MM-DD") of the study (i.e. when you want to build networks until). Please provide as the day after the last day of the study.}

\item{netwindow}{The period of time over which each network is built in months (i.e. netwindow=12 would correspond to yearly networks)}

\item{overlap}{The amount of overlap between netwindows in months (i.e. overlap=2 would result in a second network window starting 2 months before the end of the first). Overlap=0 ensures no overlap between successive network windows}

\item{nextonly}{(TRUE/FALSE). Determines whether a network edge is only created to the next capture of an individual or all captures within the intwindow. Defaults to FALSE}
}
\description{
This function creates dynamic, directed, multiplex movement networks from capture-mark-recapture datasets using information on the capture locations and times of individuals. Multiplex networks connect locations that individuals have moved between within a particular interaction window with different layers being defined by the user. The time period for each network, together with the temporal and spatial restrictions on the capture window used to infer a movement can be defined by the user
}
\examples{
data(cmr_dat2)
mindate<-"2010-01-01"
maxdate<-"2015-01-01"
intwindow<-60
netwindow<-12
overlap<-0
multimovenetdat<-MultiMoveNetCreate(data=cmr_dat,intwindow=intwindow,mindate=mindate,maxdate=maxdate,netwindow=netwindow,overlap=overlap,nextonly=TRUE)

}
