% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrData}
\alias{cmrData}
\title{cmrData}
\format{
A data frame with 1191 rows and 5 columns variables:
\itemize{
\item \code{id} - provides the identity of the individual captured
\item \code{loc} - the named capture location
\item \code{x} - geographic location of a capture event (on the horizontal axis)
\item \code{y} - geographic location of a capture event (on the vertical axis)
\item \code{date} - the capture date
}
}
\usage{
data("cmrData")
}
\description{
cmrData can be used to test social and movement network functions of CMRnet. It is an example of how dataframes should be organised to feed into CMRnet
}
\keyword{datasets}
