% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrData2}
\alias{cmrData2}
\title{cmrData2}
\format{A data frame with six variables: \code{id}, \code{loc},
  \code{x}, \code{y}, \code{date} and \code{layer}. \code{cmrData2} is an example of how dataframes should be created to feed into CMRnet when there are multiple layered movement networks (i.e. a separate network for males and females).}
\usage{
cmrData2
}
\description{
cmrData2 to test the multiplex movement network function.It contains six columns: the ID column provides the identity of the individual captured, loc is the named capture location, x and y provide the geographic location of a capture event, date provides the capture date and layer provide information used to define the layers of the multiplex network.
}
\keyword{datasets}
