% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrData3}
\alias{cmrData3}
\title{cmrData3}
\format{A data frame with five variables: \code{id}, \code{loc},
  \code{x}, \code{y} and \code{date}. \code{cmrData3} is an example of how dataframes should be created to feed into CMRnet when capture times are relevant as well as capture dates.}
\usage{
cmrData3
}
\description{
cmrData3 to test short-term social and movement network functions of CMRnet.It contains five columns: the ID column provides the identity of the individual captured, loc is the named capture location, x and y provide the geographic location of a capture event, and date provide the capture date and time.
}
\keyword{datasets}
