% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRnodeswap.R
\name{cmrNodeswap}
\alias{cmrNodeswap}
\title{cmrNodeswap}
\usage{
cmrNodeswap(cmrnet, n.rand, multi = FALSE)
}
\arguments{
\item{cmrnet}{A cmrnet object from the DynamicNetCreate, MoveNetCreate or MultiMoveNetCreate functions}

\item{n.rand}{The number of permuted matrices to be conducted by the rmperm() algorithm}

\item{multi}{(TRUE/FALSE). Indicates whether the input object is a monolayer network (from DynamicNetCreate or MoveNetCreate) or multiplex network (MultiMoveNetCreate)}
}
\description{
This function conducts network permutations of CMRnet objects, acting as a wrapper for the rmperm() function in the R package sna
Conducts node-swap randomisations for CMRnet outputs
}
\examples{
data(cmr_dat)
mindate<-"2010-01-01"
maxdate<-"2015-01-01"
intwindow<-60
netwindow<-12
overlap<-0
movenetdat<-MoveNetCreate(data=cmr_dat,intwindow=intwindow,mindate=mindate,maxdate=maxdate,netwindow=netwindow,overlap=overlap,nextonly=TRUE)
A<-cmrNodeswap(movenetdat,n.rand=1000)

data(cmr_dat2)
mindate<-"2010-01-01"
maxdate<-"2015-01-01"
intwindow<-60
netwindow<-12
overlap<-0
multimovenetdat<-MultiMoveNetCreate(data=cmr_dat,intwindow=intwindow,mindate=mindate,maxdate=maxdate,netwindow=netwindow,overlap=overlap,nextonly=TRUE)
B<-cmrNodeswap(multimovenetdat,n.rand=1000,multi=TRUE)

}
