% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmrPermSoc.R
\name{cmrPermSoc}
\alias{cmrPermSoc}
\title{cmrPermSoc}
\usage{
cmrPermSoc(D, locmat, same.time, time.restrict, same.spat, spat.restrict,
  n.swaps, n.rand, burnin, n.burnin)
}
\arguments{
\item{D}{The input dataset to be randomised}

\item{locmat}{A distance matrix between the capture locations}

\item{same.time}{(TRUE/FALSE) Whether swaps should be restricted to only occur betwen individuals trapped on the same date or not}

\item{time.restrict}{Provided as a number of months. Imposes time restrictions on when swaps can take place so that individuals can only be swapped with those a fixed time before or after being captured}

\item{same.spat}{(TRUE/FALSE) Whether swaps should be restricted to only occur between individuals trapped at the same location}

\item{spat.restrict}{Provided on the same scale as the coordinates in the input dataset. Imposes space restrictions on when swaps can take place so that individuals can only be swapped with those captued within a fixed distance}

\item{n.swaps}{The number of swaps between each random network being extracted (e.g. n.swaps = 10 would equate to 10 swaps taking place between each random network being saved)}

\item{n.rand}{The number of randomised networks to be generated}

\item{burnin}{(TRUE/FALSE) Whether burnin is required}

\item{n.burnin}{The number of swaps to discard as burn-in before the first random network is created. The total number of swaps conducted is thus n.burnin+n.swaps*n.rand}
}
\value{
A randomised dataset with the same dimensions as the original input dataset
}
\description{
An internal function that operate within CMRnet::DatastreamPermSoc
}
