% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicNetCreate.R
\name{DynamicNetCreate}
\alias{DynamicNetCreate}
\title{DynamicNetCreate}
\usage{
DynamicNetCreate(data, intwindow, mindate, maxdate, netwindow, overlap,
  spacewindow)
}
\arguments{
\item{data}{A 5 column dataframe with columns for the ID of the captured individual, the location of its capture (a name or number), the x coordinate of its capture location, the y coordinate of the capture location, and the date of capture}

\item{intwindow}{The maximum period of time (in days) between two co-captures (i.e. if \code{intwindow = 10} then two individuals captured 10 days apart could be considered co-captured but two individuals captured 11 days apart couldn't)}

\item{mindate}{The start date (format = \code{"YYYY-MM-DD"}) of the study (i.e. when you want to build networks from)}

\item{maxdate}{The end date (format = \code{"YYYY-MM-DD"}) of the study (i.e. when you want to build networks until). Please provide as the day after the last day of the study.}

\item{netwindow}{The period of time over which each network is built in months (i.e. \code{netwindow=12} would correspond to yearly networks)}

\item{overlap}{The amount of overlap between netwindows in months (i.e. \code{overlap = 2} would result in a second network window starting 2 months before the end of the first). When \code{overlap=0}, there is no overlap between successive network windows}

\item{spacewindow}{The maximum distance between locations that can be classed as a co-capture (calculated using the coordinate system provided in in the input data). Best used when multiple capture locations occur very close together}
}
\value{
A list of length 3 containing: \cr
\enumerate{
   \item The edgelist for the social network in each of the netwindows as an array
   \item The adjacency matrix for the social network in each of the netwindows as an array
   \item A matrix indicating which individuals occurred in each netwindow
}
}
\description{
This function creates dynamic social networks from capture-mark-recapture datasets using information on co-capture
The time period for each network, together with the temporal and spatial restrictions on the co-capture window can be defined by the user
}
\examples{
data(cmr_dat)
mindate<-"2010-01-01"
maxdate<-"2015-01-01"
intwindow<-60
netwindow<-12
overlap<-0
spacewindow<-0
netdat<-DynamicNetCreate(data=cmr_dat,intwindow=intwindow,mindate=mindate,maxdate=maxdate,netwindow=netwindow,overlap=overlap,spacewindow=spacewindow)

}
